+++ packages/eslint-plugin/tests/rules/no-type-alias.test.ts
@@ -443,6 +443,10 @@ type Foo<T> = {
       code: 'type MyType<T> = T extends number ? number : null;',
       options: [{ allowConditionalTypes: 'always' }],
     },
+    {
+      code: 'type Foo = new (bar: number) => string | null;',
+      options: [{ allowConstructors: 'always' }],
+    },
   ],
   invalid: [
     {
@@ -3181,6 +3185,20 @@ type Foo<T> = {
       ],
     },
     {
+      code: 'type Foo = new (bar: number) => string | null;',
+      options: [{ allowConstructors: 'never' }],
+      errors: [
+        {
+          messageId: 'noTypeAlias',
+          data: {
+            alias: 'conditional types',
+          },
+          line: 1,
+          column: 18,
+        },
+      ],
+    },
+    {
       code: 'type MyType<T> = T extends number ? number : null;',
       errors: [
         {
@@ -3190,6 +3208,10 @@ type Foo<T> = {
           },
           line: 1,
           column: 18,
+            alias: 'constructors',
+            line: 1,
+            column: 12,
+          },
         },
       ],
     },
@@ -3204,6 +3226,10 @@ type Foo<T> = {
           },
           line: 1,
           column: 18,
+            alias: 'constructors',
+            line: 1,
+            column: 12,
+          },
         },
       ],
     },